[mlir] collectSymbolScope() bail early if failed to collect references
Before this change, the `assert(limitIt->hasTrait<SymbolTable>())` fails
when there is a nested function. The function's parent is also a
function, which does not have the SymbolTable trait. This is observed in
normalize-memrefs. We can avoid this by exiting the function early if
there is a failure collecting references.

Fixes https://github.com/llvm/llvm-project/issues/61367

Differential Revision: https://reviews.llvm.org/D147399