[MLIR] Fuse all siblings that load the same memref
Fixes https://github.com/llvm/llvm-project/issues/61825

Sibling LoopFusion fuses loops that load the same memref. Prior to this
change, LoopFusion would only fuse if the those memrefs are block
arguments. This patch generalizes it to memrefs in general, which would
include memrefs that are allocated within the block.

This patch regresses @affine_2mm_fused test because it does not perform
any fusion. Prior to this change, findSiblingNodeToFuse() would target
the load of %arg0, the first block argument. After this change, we
instead target the load of %arg1, the first op in the loop. The
profitability check only determines profitability if it sees %arg0
memref (not sure why).