[ValueTracking] getUnderlyingObject wrapper to look through PHI nodes
This patch introduces getUnderlyingObjectLookThrough. It looks at the
output of getUnderlyingObject. If it is a PHI, it looks at all the
incoming underlying objects. If all those objects are the same, or the
original PHI, we determine that there is a new underlying object.

This is similar to getUnderlyingObjects, but provides a more efficient
way to find a single underlying object.

This is my attempt at solving huge compile time regressions in
https://reviews.llvm.org/D152082. First, we only look through a single
PHI, not nested PHIs. Second, we only change one callsite in BasicAA.
TODO: still seeing 3% compile time regression locally.