[IR] Support printing out uwtable attribute with none value
This avoids the pitfall where we set the uwtable to none:
```
func.setUWTableKind(llvm::UWTableKind::None)
```
Attribute::getAsString() would see an unknown attribute and fail an
assertion. In this patch, we instead print out `uwtable(none)`

This also skips the check of `UWTableKind::Async`. It is dominated by
the check of `UWTableKind::Default`, which has the same enum value.